refactor(user): userAuthInfo 네이밍 변경
- RequestDto로 두는 것보다 VO의 개념으로 두는 것이 더 올바르다고 판단.
