+++ src/main/java/com/somemore/user/dto/UserAuthInfo.java
